[MOD] Moved controller test in their folder and renamed NewMessage
