+++ src/test/java/com/chatop/api/controller/UserControllerUnitTest.java
@@ -1,4 +1,4 @@
-package com.chatop.api;
+package com.chatop.api.controller;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,7 +9,6 @@ import org.springframework.test.web.servlet.MockMvc;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.chatop.api.controller.UserController;
 import com.chatop.api.service.UserServiceImpl;
 
 @WebMvcTest(controllers = UserController.class)
